test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

hacking>=0.9.2,<0.10
# mock object framework
mock>=1.0
coverage>=3.6
discover
# fixture stubbing
fixtures>=0.3.14
oslotest>=1.1.0  # Apache-2.0
python-subunit
junitxml
nose
nose-exclude
nosexcover
# Doc requirements
sphinx>=1.1.2,<1.1.999
sphinxcontrib-pecanwsme>=0.6,<=0.8.0
oslosphinx<=2.5.0
